rename process-eye to image-flop
spoof of photoshop!
