+++ image-flop/INSTRUCTIONS_IMAGEFLOP.md
@@ -1,4 +1,4 @@
-# Process-Eye
+# Image-flop
 A simple image processing program built in QML.
 
 ## Requirements
