+++ image-flop/include/imageprocessor.h
