+++ image-flop/qml/FilterModel.qml
