--- process-eye/resources.qrc
@@ -1,11 +1,0 @@
-<!DOCTYPE RCC>
-<RCC version="1.0">
-
-<qresource prefix="/">
-    <file>qml/process-eye.qml</file>
-	<file>qml/MainContent.qml</file>
-	<file>qml/FilterModel.qml</file>
-    <file>eye.jpg</file>
-</qresource>
-
-</RCC>
